# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Item": {
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"price",
					"description",
					"image"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"id": 1,
					"name": "My item",
					"price": 10.5,
					"description": "My item description",
					"image": "http://image"
				}
			},
			"ValidationProblemDetails": {
				"properties": {
					"type": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"status": {
						"type": "integer",
						"format": "int32"
					},
					"detail": {
						"type": "string"
					},
					"instance": {
						"type": "string"
					},
					"errors": {
						"properties": {},
						"additionalProperties": {
							"properties": {
								"value": {},
								"message": {
									"type": "string"
								}
							},
							"required": [
								"message"
							],
							"type": "object"
						},
						"type": "object"
					}
				},
				"required": [
					"type",
					"title",
					"status",
					"errors"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"type": "https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.1",
					"title": "One or more validation errors occurred",
					"status": 400,
					"errors": {
						"id": {
							"message": "id",
							"value": "invalid"
						}
					}
				}
			},
			"ProblemDetails": {
				"properties": {
					"type": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"status": {
						"type": "integer",
						"format": "int32"
					},
					"detail": {
						"type": "string"
					},
					"instance": {
						"type": "string"
					}
				},
				"required": [
					"type",
					"title",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "k8s-node-api",
		"version": "1.0.0",
		"description": "A demo API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/items/{id}": {
			"get": {
				"operationId": "GetItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Item"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": [
					"items"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}